+++ medusa/model/modeling_mistral_kv.py
@@ -1080,7 +1080,7 @@ class MistralForCausalLM(MistralPreTrainedModel):
         )
 
         hidden_states = outputs[0]
-        logits = self.lm_head(hidden_states)
+        logits = self.lm_head(hidden_states.to(self.lm_head.weight.dtype))
         logits = logits.float()
 
         loss = None
